Safari-609.1.11.1/LayoutTests/fast/events/input-events-drag-and-drop-expected.txt

To manually test this, drag the text in the first input into the second, then undo (cmd-z) and observe the output. You should see two deleteByDrag events (beforeinput and input) followed by two insertFromDrop events. After undoing, you should see two historyUndo events, and redoing should output two historyRedo events.


Performing drag and drop
(source): type=beforeinput, inputType=deleteByDrag, data=`null`
(destination): type=beforeinput, inputType=insertFromDrop, data=`Input events!`
(source): type=input, inputType=deleteByDrag, data=`null`
(destination): type=input, inputType=insertFromDrop, data=`Input events!`

Undoing drag and drop
(destination): type=beforeinput, inputType=historyUndo, data=`null`
(destination): type=input, inputType=historyUndo, data=`null`

Redoing drag and drop
(destination): type=beforeinput, inputType=historyRedo, data=`null`
(destination): type=input, inputType=historyRedo, data=`null`
